requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.11.1
    # via
    #   django
    #   django-allauth
    #   django-axes
    #   django-cors-headers
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
bandit==1.9.4
    # via -r requirements.in
billiard==4.2.4
    # via celery
boolean-py==5.0
    # via license-expression
brotli==1.2.0
    # via fonttools
cachecontrol[filecache]==0.14.4
    # via
    #   cachecontrol
    #   pip-audit
celery==5.6.3
    # via -r requirements.in
certifi==2026.2.25
    # via
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via weasyprint
cfgv==3.5.0
    # via pre-commit
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
coverage==7.13.5
    # via -r requirements.in
cssselect2==0.9.0
    # via weasyprint
cyclonedx-python-lib==11.7.0
    # via pip-audit
defusedxml==0.7.1
    # via py-serializable
deprecated==1.3.1
    # via pikepdf
distlib==0.4.0
    # via virtualenv
django==5.2.13
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-axes
    #   django-cors-headers
    #   django-csp
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-prometheus
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth==65.16.1
    # via -r requirements.in
django-axes==8.3.1
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
django-csp==4.0
    # via -r requirements.in
django-debug-toolbar==6.3.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-filter==25.2
    # via -r requirements.in
django-prometheus==2.4.1
    # via -r requirements.in
django-redis==6.0.0
    # via -r requirements.in
djangorestframework==3.17.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-spectacular==0.29.0
    # via -r requirements.in
elementpath==5.1.1
    # via xmlschema
environs==15.0.1
    # via -r requirements.in
factory-boy==3.3.3
    # via -r requirements.in
factur-x==4.2
    # via -r requirements.in
faker==40.15.0
    # via factory-boy
filelock==3.29.0
    # via
    #   cachecontrol
    #   python-discovery
    #   virtualenv
fonttools[woff]==4.62.1
    # via weasyprint
gunicorn==25.3.0
    # via -r requirements.in
identify==2.6.19
    # via pre-commit
idna==3.11
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.3.0
    # via pytest
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
kombu==5.6.2
    # via celery
license-expression==30.4.4
    # via cyclonedx-python-lib
lxml==6.1.0
    # via
    #   -r requirements.in
    #   factur-x
    #   pikepdf
markdown-it-py==4.0.0
    # via rich
marshmallow==4.3.0
    # via environs
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.2
    # via cachecontrol
nodeenv==1.10.0
    # via pre-commit
packageurl-python==0.17.6
    # via cyclonedx-python-lib
packaging==26.1
    # via
    #   django-csp
    #   gunicorn
    #   kombu
    #   pikepdf
    #   pip-audit
    #   pip-requirements-parser
    #   pytest
    #   wheel
pikepdf==10.5.1
    # via -r requirements.in
pillow==12.2.0
    # via
    #   -r requirements.in
    #   pikepdf
    #   weasyprint
pip-api==0.0.34
    # via pip-audit
pip-audit==2.10.0
    # via -r requirements.in
pip-requirements-parser==32.0.1
    # via pip-audit
platformdirs==4.9.6
    # via
    #   pip-audit
    #   python-discovery
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.5.1
    # via -r requirements.in
prometheus-client==0.25.0
    # via
    #   -r requirements.in
    #   django-prometheus
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements.in
py-serializable==2.1.0
    # via cyclonedx-python-lib
pycparser==3.0
    # via cffi
pydyf==0.12.1
    # via weasyprint
pygments==2.20.0
    # via
    #   pytest
    #   rich
pyjwt==2.12.1
    # via djangorestframework-simplejwt
pyparsing==3.3.2
    # via pip-requirements-parser
pypdf==6.10.2
    # via factur-x
pyphen==0.17.2
    # via weasyprint
pytest==9.0.3
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.12.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-discovery==1.2.2
    # via virtualenv
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   environs
python-json-logger==4.1.0
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   bandit
    #   drf-spectacular
    #   pre-commit
redis==7.4.0
    # via
    #   -r requirements.in
    #   django-redis
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via
    #   cachecontrol
    #   pip-audit
rich==15.0.0
    # via
    #   bandit
    #   pip-audit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
ruff==0.15.11
    # via -r requirements.in
saxonche==12.9.0
    # via
    #   -r requirements.in
    #   factur-x
sentry-sdk==2.58.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sortedcontainers==2.4.0
    # via cyclonedx-python-lib
sqlparse==0.5.5
    # via
    #   django
    #   django-debug-toolbar
stevedore==5.7.0
    # via bandit
tinycss2==1.5.1
    # via
    #   cssselect2
    #   weasyprint
tinyhtml5==2.1.0
    # via weasyprint
tomli==2.4.1
    # via pip-audit
tomli-w==1.2.0
    # via pip-audit
tzdata==2026.1
    # via kombu
tzlocal==5.3.1
    # via celery
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.6.3
    # via
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==21.2.4
    # via pre-commit
wcwidth==0.6.0
    # via prompt-toolkit
weasyprint==68.1
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
    #   tinyhtml5
wheel==0.46.3
    # via -r requirements.in
whitenoise==6.12.0
    # via -r requirements.in
wrapt==2.1.2
    # via deprecated
xmlschema==4.3.1
    # via -r requirements.in
zopfli==0.4.1
    # via fonttools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
